CAMEL-16400: split unit and integration tests for camel-aws2-kms
