+++ components/camel-aws/camel-aws2-kms/src/test/java/org/apache/camel/component/aws2/kms/localstack/KmsDisableKeyIT.java
@@ -25,12 +25,12 @@ import org.apache.camel.component.aws2.kms.KMS2Constants;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.junit.jupiter.api.Test;
 import software.amazon.awssdk.services.kms.model.CreateKeyResponse;
-import software.amazon.awssdk.services.kms.model.DescribeKeyResponse;
-import software.amazon.awssdk.services.kms.model.KeyState;
+import software.amazon.awssdk.services.kms.model.ListKeysResponse;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 
-public class KmsScheduleDeleteKeyLocalstackTest extends Aws2KmsBaseTest {
+public class KmsDisableKeyIT extends Aws2KmsBase {
 
     @EndpointInject
     private ProducerTemplate template;
@@ -52,28 +52,26 @@ public class KmsScheduleDeleteKeyLocalstackTest extends Aws2KmsBaseTest {
 
         String keyId = ex.getMessage().getBody(CreateKeyResponse.class).keyMetadata().keyId();
 
-        template.send("direct:scheduleDelete", new Processor() {
+        template.send("direct:disableKey", new Processor() {
 
             @Override
             public void process(Exchange exchange) throws Exception {
-                exchange.getIn().setHeader(KMS2Constants.OPERATION, "scheduleKeyDeletion");
+                exchange.getIn().setHeader(KMS2Constants.OPERATION, "disableKey");
                 exchange.getIn().setHeader(KMS2Constants.KEY_ID, keyId);
             }
         });
 
-        template.send("direct:describeKey", new Processor() {
+        template.send("direct:listKeys", new Processor() {
 
             @Override
             public void process(Exchange exchange) throws Exception {
-                exchange.getIn().setHeader(KMS2Constants.OPERATION, "describeKey");
-                exchange.getIn().setHeader(KMS2Constants.KEY_ID, keyId);
+                exchange.getIn().setHeader(KMS2Constants.OPERATION, "listKeys");
             }
         });
 
         assertMockEndpointsSatisfied();
         assertEquals(1, result.getExchanges().size());
-        assertEquals(KeyState.PENDING_DELETION,
-                result.getExchanges().get(0).getIn().getBody(DescribeKeyResponse.class).keyMetadata().keyState());
+        assertTrue(result.getExchanges().get(0).getIn().getBody(ListKeysResponse.class).hasKeys());
     }
 
     @Override
@@ -83,13 +81,11 @@ public class KmsScheduleDeleteKeyLocalstackTest extends Aws2KmsBaseTest {
             public void configure() throws Exception {
                 String awsEndpoint
                         = "aws2-kms://default?operation=createKey";
-                String describeKey
-                        = "aws2-kms://default?operation=describeKey";
-                String scheduleDelete
-                        = "aws2-kms://default?operation=describeKey";
+                String disableKey = "aws2-kms://default?operation=disableKey";
+                String listKeys = "aws2-kms://default?operation=listKeys";
                 from("direct:createKey").to(awsEndpoint);
-                from("direct:scheduleDelete").to(scheduleDelete);
-                from("direct:describeKey").to(describeKey).to("mock:result");
+                from("direct:disableKey").to(disableKey);
+                from("direct:listKeys").to(listKeys).to("mock:result");
             }
         };
     }
