+++ components/camel-aws/camel-aws2-kms/src/test/java/org/apache/camel/component/aws2/kms/localstack/KmsListKeysIT.java
@@ -24,13 +24,12 @@ import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.aws2.kms.KMS2Constants;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.junit.jupiter.api.Test;
-import software.amazon.awssdk.services.kms.model.CreateKeyResponse;
-import software.amazon.awssdk.services.kms.model.DescribeKeyResponse;
-import software.amazon.awssdk.services.kms.model.KeyState;
+import software.amazon.awssdk.services.kms.model.ListKeysResponse;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 
-public class KmsDescribeKeyLocalstackTest extends Aws2KmsBaseTest {
+public class KmsListKeysIT extends Aws2KmsBase {
 
     @EndpointInject
     private ProducerTemplate template;
@@ -42,7 +41,7 @@ public class KmsDescribeKeyLocalstackTest extends Aws2KmsBaseTest {
     public void sendIn() throws Exception {
         result.expectedMessageCount(1);
 
-        Exchange ex = template.send("direct:createKey", new Processor() {
+        template.send("direct:createKey", new Processor() {
 
             @Override
             public void process(Exchange exchange) throws Exception {
@@ -50,21 +49,17 @@ public class KmsDescribeKeyLocalstackTest extends Aws2KmsBaseTest {
             }
         });
 
-        String keyId = ex.getMessage().getBody(CreateKeyResponse.class).keyMetadata().keyId();
-
-        template.send("direct:describeKey", new Processor() {
+        template.send("direct:listKeys", new Processor() {
 
             @Override
             public void process(Exchange exchange) throws Exception {
-                exchange.getIn().setHeader(KMS2Constants.OPERATION, "describeKey");
-                exchange.getIn().setHeader(KMS2Constants.KEY_ID, keyId);
+                exchange.getIn().setHeader(KMS2Constants.OPERATION, "listKeys");
             }
         });
 
         assertMockEndpointsSatisfied();
         assertEquals(1, result.getExchanges().size());
-        assertEquals(KeyState.ENABLED,
-                result.getExchanges().get(0).getIn().getBody(DescribeKeyResponse.class).keyMetadata().keyState());
+        assertTrue(result.getExchanges().get(0).getIn().getBody(ListKeysResponse.class).hasKeys());
     }
 
     @Override
@@ -74,10 +69,9 @@ public class KmsDescribeKeyLocalstackTest extends Aws2KmsBaseTest {
             public void configure() throws Exception {
                 String awsEndpoint
                         = "aws2-kms://default?operation=createKey";
-                String describeKey
-                        = "aws2-kms://default?operation=describeKey";
+                String listKeys = "aws2-kms://default?operation=listKeys";
                 from("direct:createKey").to(awsEndpoint);
-                from("direct:describeKey").to(describeKey).to("mock:result");
+                from("direct:listKeys").to(listKeys).to("mock:result");
             }
         };
     }
